Add Normalizer class for field value normalization
