+++ src/Validator.php
@@ -6,30 +6,6 @@ namespace EForms;
 class Validator
 {
     /**
-     * Registry mapping normalizer IDs to callable handlers.
-     */
-    private const HANDLERS = [
-        '' => [self::class, 'identity'],
-        'text' => [self::class, 'identity'],
-        'email' => [self::class, 'normalizeEmail'],
-        'url' => [self::class, 'identity'],
-        'tel' => [self::class, 'identity'],
-        'tel_us' => [self::class, 'normalizeTelUs'],
-        'number' => [self::class, 'identity'],
-        'range' => [self::class, 'identity'],
-        'date' => [self::class, 'identity'],
-        'textarea' => [self::class, 'identity'],
-        'textarea_html' => [self::class, 'identity'],
-        'zip' => [self::class, 'identity'],
-        'zip_us' => [self::class, 'identity'],
-        'select' => [self::class, 'identity'],
-        'radio' => [self::class, 'identity'],
-        'checkbox' => [self::class, 'identity'],
-        'file' => [self::class, 'identity'],
-        'files' => [self::class, 'identity'],
-    ];
-
-    /**
      * Registry mapping validator IDs to callable handlers.
      */
     private const VALIDATORS = [
@@ -54,55 +30,16 @@ class Validator
     ];
 
     /**
-     * Resolve a handler by identifier.
+     * Resolve a validator handler by identifier.
      *
-     * @param string $kind Either 'normalizer' or 'validator'
      * @throws \RuntimeException when the identifier is unknown
      */
-    public static function resolve(string $id, string $kind = 'normalizer'): callable
+    public static function resolve(string $id): callable
     {
-        $map = $kind === 'validator' ? self::VALIDATORS : self::HANDLERS;
-        if (!isset($map[$id])) {
-            throw new \RuntimeException('Unknown ' . $kind . ' ID: ' . $id);
+        if (!isset(self::VALIDATORS[$id])) {
+            throw new \RuntimeException('Unknown validator ID: ' . $id);
         }
-        return $map[$id];
-    }
-
-    /**
-     * Default passthrough handler used for types that do not require
-     * additional normalization.
-     *
-     * @param mixed $v
-     * @return mixed
-     */
-    public static function identity($v)
-    {
-        return $v;
-    }
-
-    /**
-     * Normalize email addresses by lowercasing the domain component.
-     */
-    public static function normalizeEmail(string $v): string
-    {
-        if ($v !== '' && strpos($v, '@') !== false) {
-            [$local, $domain] = explode('@', $v, 2);
-            return $local . '@' . strtolower($domain);
-        }
-        return $v;
-    }
-
-    /**
-     * Normalize US telephone numbers by stripping non-digits and
-     * removing a leading country code of 1.
-     */
-    public static function normalizeTelUs(string $v): string
-    {
-        $digits = preg_replace('/\D+/', '', $v);
-        if (strlen($digits) === 11 && str_starts_with($digits, '1')) {
-            $digits = substr($digits, 1);
-        }
-        return $digits;
+        return self::VALIDATORS[$id];
     }
 
     /**
@@ -283,8 +220,8 @@ class Validator
             $nid = $f['handlers']['normalizer_id'] ?? ($f['type'] ?? '');
             $f['form_id'] = $tpl['id'] ?? '';
             $f['handlers'] = [
-                'validator'  => self::resolve($hid, 'validator'),
-                'normalizer' => self::resolve($nid, 'normalizer'),
+                'validator'  => self::resolve($hid),
+                'normalizer' => Normalizer::resolve($nid),
             ];
             $desc[$f['key']] = $f;
         }
@@ -296,7 +233,7 @@ class Validator
         $desc = $desc ?? self::descriptors($tpl);
         $values = [];
         foreach ($desc as $k => $f) {
-            $norm = $f['handlers']['normalizer'] ?? [self::class, 'identity'];
+            $norm = $f['handlers']['normalizer'] ?? [Normalizer::class, 'identity'];
             if (self::isMultivalue($f)) {
                 $raw = $post[$k] ?? [];
                 if (!is_array($raw)) {
@@ -384,7 +321,7 @@ class Validator
     {
         $out = [];
         foreach ($desc as $k => $f) {
-            $norm = $f['handlers']['normalizer'] ?? [self::class, 'identity'];
+            $norm = $f['handlers']['normalizer'] ?? [Normalizer::class, 'identity'];
             $v = $values[$k] ?? (self::isMultivalue($f) ? [] : '');
             if (self::isMultivalue($f)) {
                 $out[$k] = array_map($norm, is_array($v) ? $v : []);
