+++ tests/DescriptorsResolutionTest.php
@@ -5,6 +5,7 @@ use PHPUnit\Framework\TestCase;
 use EForms\Spec;
 use EForms\Validator;
 use EForms\Renderer;
+use EForms\Normalizer;
 
 final class DescriptorsResolutionTest extends TestCase
 {
@@ -15,8 +16,8 @@ final class DescriptorsResolutionTest extends TestCase
             $this->assertArrayHasKey('type', $desc);
             $type = $desc['type'];
             $handlers = $desc['handlers'];
-            $v = Validator::resolve($handlers['validator_id'] ?? '', 'validator');
-            $n = Validator::resolve($handlers['normalizer_id'] ?? '', 'normalizer');
+            $v = Validator::resolve($handlers['validator_id'] ?? '');
+            $n = Normalizer::resolve($handlers['normalizer_id'] ?? '');
             $r = Renderer::resolve($handlers['renderer_id'] ?? '');
             $this->assertTrue(is_callable($v), "$type validator");
             $this->assertTrue(is_callable($n), "$type normalizer");
@@ -36,6 +37,6 @@ final class DescriptorsResolutionTest extends TestCase
     public function testUnknownHandlerIdFails(): void
     {
         $this->expectException(\RuntimeException::class);
-        Validator::resolve('unknown', 'validator');
+        Validator::resolve('unknown');
     }
 }
